Вынес в game-*.js часть логики + возврат HTML-элемента
